Move tests for better organization
